Allow choice of shift_mode or manually set shifts in Nscf. Update for new (in new abipy) factory functions for nscf/dos calculations.
